fill out bits of docking function
